+++ pkg/controllers/dynakube/oneagent/oneagent_reconciler.go
@@ -22,7 +22,6 @@ import (
 	"github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/oneagent/daemonset"
 	"github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/token"
 	"github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/version"
-	"github.com/Dynatrace/dynatrace-operator/pkg/injection/namespace/bootstrapperconfig"
 	"github.com/Dynatrace/dynatrace-operator/pkg/logd"
 	"github.com/Dynatrace/dynatrace-operator/pkg/util/hasher"
 	"github.com/Dynatrace/dynatrace-operator/pkg/util/kubernetes/fields/k8sconditions"
@@ -311,19 +310,6 @@ func (r *Reconciler) getOneagentPods(ctx context.Context, dk *dynakube.DynaKube,
 }
 
 func (r *Reconciler) buildDesiredDaemonSet(ctx context.Context, dk *dynakube.DynaKube) (*appsv1.DaemonSet, error) {
-	var configHash string
-
-	if bootstrapperconfig.NeedsPGC(dk) {
-		var err error
-
-		configHash, err = r.pgcConfigHash(ctx, dk)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	pgcReady := configHash != ""
-
 	var ds *appsv1.DaemonSet
 
 	var err error
@@ -332,17 +318,13 @@ func (r *Reconciler) buildDesiredDaemonSet(ctx context.Context, dk *dynakube.Dyn
 	case dk.OneAgent().IsClassicFullStackMode():
 		ds, err = daemonset.NewClassicFullStack(dk, r.clusterID).BuildDaemonSet(ctx)
 	case dk.OneAgent().IsHostMonitoringMode():
-		ds, err = daemonset.NewHostMonitoring(dk, r.clusterID, pgcReady).BuildDaemonSet(ctx)
+		ds, err = daemonset.NewHostMonitoring(dk, r.clusterID, r.apiReader).BuildDaemonSet(ctx)
 	case dk.OneAgent().IsCloudNativeFullstackMode():
-		ds, err = daemonset.NewCloudNativeFullStack(dk, r.clusterID, pgcReady).BuildDaemonSet(ctx)
+		ds, err = daemonset.NewCloudNativeFullStack(dk, r.clusterID, r.apiReader).BuildDaemonSet(ctx)
 	}
 
 	if err != nil {
 		return nil, err
-	}
-
-	if pgcReady {
-		ds.Spec.Template.Annotations[daemonset.AnnotationPGCHash] = configHash
 	}
 
 	dsHash, err := hasher.GenerateHash(ds)
@@ -353,26 +335,6 @@ func (r *Reconciler) buildDesiredDaemonSet(ctx context.Context, dk *dynakube.Dyn
 	ds.Annotations[hasher.AnnotationHash] = dsHash
 
 	return ds, nil
-}
-
-func (r *Reconciler) pgcConfigHash(ctx context.Context, dk *dynakube.DynaKube) (string, error) {
-	var secret corev1.Secret
-
-	err := r.apiReader.Get(ctx, client.ObjectKey{Name: bootstrapperconfig.GetSourceConfigSecretName(dk.Name), Namespace: dk.Namespace}, &secret)
-	if k8serrors.IsNotFound(err) {
-		return "", nil
-	}
-
-	if err != nil {
-		return "", errors.WithStack(err)
-	}
-
-	pgcData := secret.Data[bootstrapperconfig.DeclarativeInputFileName]
-	if pgcData == nil {
-		return "", nil
-	}
-
-	return hasher.GenerateHash(pgcData)
 }
 
 func (r *Reconciler) reconcileInstanceStatuses(ctx context.Context, dk *dynakube.DynaKube) error {
